requirements.txt
# Django Core
Django==5.0.1
django-environ==0.11.2
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9

# Web Scraping & HTTP
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
selenium==4.16.0
playwright==1.40.0
urllib3==2.1.0

# AI & ML
google-generativeai==0.3.2
openai==1.6.1
langchain==0.1.0
langchain-google-genai==0.0.6

# Data Processing
pandas==2.1.4
numpy==1.26.3
openpyxl==3.1.2

# Image Processing
Pillow==10.2.0

# Forms & Validation
django-crispy-forms==2.1
crispy-bootstrap5==2.0.0

# Security & Authentication
django-allauth==0.61.1
PyJWT==2.8.0
cryptography==42.0.0

# API & REST
djangorestframework==3.14.0
django-cors-headers==4.3.1

# File Handling
python-magic==0.4.27
python-magic-bin==0.4.14

# Date & Time
python-dateutil==2.8.2
pytz==2024.1

# Celery (Background Tasks)
celery==5.3.6
redis==5.0.1

# Email
django-anymail==10.2

# Payment Integration (Optional)
stripe==7.9.0
razorpay==1.4.1

# Maps & Location
geopy==2.4.1
googlemaps==4.10.0

# Charts & Visualization
plotly==5.18.0
matplotlib==3.8.2

# PDF Generation
reportlab==4.0.9
weasyprint==60.2

# QR Code
qrcode==7.4.2
python-barcode==0.15.1

# CSV & Excel
xlsxwriter==3.1.9

# Caching
django-redis==5.4.0

# Development Tools
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Testing
pytest==7.4.4
pytest-django==4.7.0
faker==22.0.0

# Code Quality
flake8==7.0.0
black==23.12.1
isort==5.13.2

# Deployment
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0

# Monitoring
sentry-sdk==1.40.0

# Utilities
python-slugify==8.0.1
shortuuid==1.0.11
